add convert.go
